% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labmkt}
\alias{labmkt}
\title{Statistics New Zealand Labour Market Time Series (March 2016)}
\format{A data table with columns:
\describe{
  \item{ref}{Series reference.}
  \item{period}{String describing date.}
  \item{value}{Time series value.}
  \item{status}{Status--final, provisional, or revised.}
}}
\usage{
labmkt
}
\description{
Time series data for Statistics New Zealand's Household Labour Force Survey (HLFS), Quarterly Employment Survey (QES), and Labour Cost Index (LCI).
}
\keyword{datasets}

