% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesdemo.R
\name{timeseriesdemo}
\alias{timeseriesdemo}
\title{Interface for Statistics New Zealand Labour Market Time Series Data.}
\usage{
timeseriesdemo()
}
\description{
Simple shiny application to explore Statistics New Zealand labour market time series data.
}
\examples{
\dontrun{
}
}
\author{
Chris Hansen
}

