% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesdemo2.R
\name{timeseriesdemo2}
\alias{timeseriesdemo2}
\title{Interface for Statistics New Zealand Labour Market Time Series Data.}
\usage{
timeseriesdemo2()
}
\description{
Simple shiny application to explore Statistics New Zealand labour market time series data.
}
\examples{
\dontrun{
}
}
\author{
Chris Hansen
}

